+++ yash-syntax/src/parser.rs
@@ -38,8 +38,10 @@
 //! let input = Box::new(Memory::new("echo $?", Source::Unknown));
 //!
 //! // Next, create a lexer.
+//! # use std::num::NonZeroU64;
 //! # use yash_syntax::parser::lex::Lexer;
-//! let mut lexer = Lexer::new(input);
+//! let line = NonZeroU64::new(1).unwrap();
+//! let mut lexer = Lexer::new(input, line, Source::Unknown);
 //!
 //! // Then, create a new parser borrowing the lexer.
 //! # use yash_syntax::parser::Parser;